apischema: Move PackageDetails to apischema
And rename it to PackageDetailsV1, following the same pattern as
weldr/apischema.go

Related: RHEL-60147
