+++ cloud/search.go
@@ -3,34 +3,7 @@ package cloud
 import (
 	"encoding/json"
 	"fmt"
-
-	"github.com/osbuild/weldr-client/v2/internal/common"
 )
-
-type PackageDetails struct {
-	common.PackageNEVRA
-	detailFields
-}
-
-type detailFields struct {
-	Summary     string `json:"summary"`
-	Description string `json:"description"`
-	Buildtime   string `json:"buildtime"`
-	License     string `json:"license"`
-	URL         string `json:"url"`
-}
-
-func (pkg *PackageDetails) UnmarshalJSON(data []byte) error {
-	if err := json.Unmarshal(data, &pkg.PackageNEVRA); err != nil {
-		return err
-	}
-
-	if err := json.Unmarshal(data, &pkg.detailFields); err != nil {
-		return err
-	}
-
-	return nil
-}
 
 type searchRequest struct {
 	Distribution string   `json:"distribution"`
@@ -40,7 +13,7 @@ type searchRequest struct {
 
 // SearchPackages returns details about the packages
 // Wildcards are supported with '*'
-func (c Client) SearchPackages(packages []string, distro, arch string) ([]PackageDetails, error) {
+func (c Client) SearchPackages(packages []string, distro, arch string) ([]PackageDetailsV1, error) {
 	request := searchRequest{
 		Distribution: distro,
 		Architecture: arch,
@@ -56,7 +29,7 @@ func (c Client) SearchPackages(packages []string, distro, arch string) ([]Packag
 	}
 
 	var r struct {
-		Packages []PackageDetails `json:"packages"`
+		Packages []PackageDetailsV1 `json:"packages"`
 	}
 	err = json.Unmarshal(body, &r)
 	if err != nil {
